Move the new data toolbar to the components directory
https://github.com/project-koku/koku-ui/issues/1275
