+++ src/api/query.ts
@@ -1,0 +1,9 @@
+type FilterByValue = string | string[];
+
+interface FilterBys {
+  tag?: FilterByValue;
+}
+
+export interface Query {
+  filter_by?: FilterBys;
+}
