+++ src/utils/getComputedOcpCloudReportItems.ts
@@ -173,8 +173,5 @@ export function getIdKeyForGroupBy(
   if (groupBy.node) {
     return 'node';
   }
-  if (groupBy.tags) {
-    return 'tags';
-  }
   return 'date';
 }
